Updates to testing for nomenclate.  Realizing I need to refactor a lot of this code, but try to keep the API the same for currently existing scripts in other projects.
Removed set command since it was the same and worse than reset...might have to rename reset for a more obvious name since it doesn't only reset but can set to other values, maybe a flag?

Need to completely refactor how the configurator pulls keys.  Too much is hardcoded for specific formats, this should be a much broader algorithm since currently it's too specialized.  If a user needs a more custom setup this could get hairy, so I need to have them able to inherit and set up other functions...or something...need to think on this and what design pattern would be best case for extensibility.  Also changed get_state to return a state dictionary with hash tables.  It might be easier to override the internal equals function instead of returning state since we already have get_dict which does a similar thing and users don't need to know hash values for any reason I can think of...